+++ etc/run-clang-tidy.sh
@@ -28,7 +28,7 @@ cmake_config_flags=(
   -DCMAKE_EXPORT_COMPILE_COMMANDS=ON
   -DCMAKE_BUILD_TYPE=Debug
   -DCMAKE_PREFIX_PATH="$(pwd)/../mongoc" # Avoid downloading C Driver.
-  -DCMAKE_CXX_STANDARD=17                # Avoid downloading mnmlstc/core.
+  -DCMAKE_CXX_STANDARD=17
 )
 
 # Generate the compilation database file.