+++ src/bsoncxx/include/bsoncxx/v_noabi/bsoncxx/stdx/string_view.hpp
@@ -16,74 +16,7 @@
 
 #include <bsoncxx/config/prelude.hpp>
 
-#if defined(BSONCXX_POLY_USE_MNMLSTC)
-
-#include <core/string.hpp>
-
-namespace bsoncxx {
-namespace v_noabi {
-namespace stdx {
-
-using ::core::basic_string_view;
-using ::core::string_view;
-
-}  // namespace stdx
-}  // namespace v_noabi
-}  // namespace bsoncxx
-
-#elif defined(BSONCXX_POLY_USE_BOOST)
-
-#include <boost/version.hpp>
-
-#if BOOST_VERSION >= 106100
-
-#include <boost/utility/string_view.hpp>
-
-namespace bsoncxx {
-namespace v_noabi {
-namespace stdx {
-
-using ::boost::basic_string_view;
-using ::boost::string_view;
-
-}  // namespace stdx
-}  // namespace v_noabi
-}  // namespace bsoncxx
-
-#else
-
-#include <boost/utility/string_ref.hpp>
-
-namespace bsoncxx {
-namespace v_noabi {
-namespace stdx {
-
-template <typename charT, typename traits = std::char_traits<charT>>
-using basic_string_view = ::boost::basic_string_ref<charT, traits>;
-using string_view = ::boost::string_ref;
-
-}  // namespace stdx
-}  // namespace v_noabi
-}  // namespace bsoncxx
-
-#endif
-
-#elif defined(BSONCXX_POLY_USE_STD_EXPERIMENTAL)
-
-#include <experimental/string_view>
-
-namespace bsoncxx {
-namespace v_noabi {
-namespace stdx {
-
-using ::std::experimental::basic_string_view;
-using ::std::experimental::string_view;
-
-}  // namespace stdx
-}  // namespace v_noabi
-}  // namespace bsoncxx
-
-#elif defined(BSONCXX_POLY_USE_STD)
+#if defined(BSONCXX_POLY_USE_STD)
 
 #include <string_view>